+++ include/memblock.h
@@ -10,50 +10,50 @@
 
 
 
-struct ag_mblock_exception {
+struct ag_memblock_exception {
         size_t sz;
         size_t align;
 };
 
 
-extern void ag_mblock_exception_handler(const struct ag_exception *, void *);
+extern void ag_memblock_exception_handler(const struct ag_exception *, void *);
 
-typedef void ag_mblock;
+typedef void ag_memblock;
 
-extern ag_mblock *ag_mblock_new(size_t);
-extern ag_mblock *ag_mblock_new_align(size_t, size_t);
-extern ag_mblock *ag_mblock_copy(const ag_mblock *);
-extern ag_mblock *ag_mblock_clone(const ag_mblock *);
-extern ag_mblock *ag_mblock_clone_align(const ag_mblock *, size_t);
-extern void ag_mblock_release(ag_mblock **);
+extern ag_memblock *ag_memblock_new(size_t);
+extern ag_memblock *ag_memblock_new_align(size_t, size_t);
+extern ag_memblock *ag_memblock_copy(const ag_memblock *);
+extern ag_memblock *ag_memblock_clone(const ag_memblock *);
+extern ag_memblock *ag_memblock_clone_align(const ag_memblock *, size_t);
+extern void ag_memblock_release(ag_memblock **);
 
 // warning: don't use with structs containing non-scalar members
 // if not equal, comparison based on first differing byte
-extern enum ag_cmp ag_mblock_cmp(const ag_mblock *, const ag_mblock *cmp);
+extern enum ag_cmp ag_memblock_cmp(const ag_memblock *, const ag_memblock *cmp);
 
-inline bool ag_mblock_lt(const ag_mblock *ctx, const ag_mblock *cmp)
+inline bool ag_memblock_lt(const ag_memblock *ctx, const ag_memblock *cmp)
 {
-        return ag_mblock_cmp(ctx, cmp) == AG_CMP_LT;
+        return ag_memblock_cmp(ctx, cmp) == AG_CMP_LT;
 }
 
-inline bool ag_mblock_eq(const ag_mblock *ctx, const ag_mblock *cmp)
+inline bool ag_memblock_eq(const ag_memblock *ctx, const ag_memblock *cmp)
 {
-        return ag_mblock_cmp(ctx, cmp) == AG_CMP_EQ;
+        return ag_memblock_cmp(ctx, cmp) == AG_CMP_EQ;
 }
 
-inline bool ag_mblock_gt(const ag_mblock *ctx, const ag_mblock *cmp)
+inline bool ag_memblock_gt(const ag_memblock *ctx, const ag_memblock *cmp)
 {
-        return ag_mblock_cmp(ctx, cmp) == AG_CMP_GT;
+        return ag_memblock_cmp(ctx, cmp) == AG_CMP_GT;
 }
 
-extern size_t ag_mblock_sz(const ag_mblock *);
-extern size_t ag_mblock_sz_total(const ag_mblock *);
-extern size_t ag_mblock_refc(const ag_mblock *);
-extern bool ag_mblock_aligned(const ag_mblock *, size_t);
+extern size_t ag_memblock_sz(const ag_memblock *);
+extern size_t ag_memblock_sz_total(const ag_memblock *);
+extern size_t ag_memblock_refc(const ag_memblock *);
+extern bool ag_memblock_aligned(const ag_memblock *, size_t);
 
-extern void ag_mblock_resize(ag_mblock **, size_t);
-extern void ag_mblock_resize_align(ag_mblock **, size_t, size_t);
-extern ag_string *ag_mblock_str(const ag_mblock *);
+extern void ag_memblock_resize(ag_memblock **, size_t);
+extern void ag_memblock_resize_align(ag_memblock **, size_t, size_t);
+extern ag_string *ag_memblock_str(const ag_memblock *);
 
 
 #ifdef __cplusplus
