+++ src/memblock.c
@@ -13,28 +13,28 @@
 #endif
 
 
-static inline size_t *meta_head(const ag_mblock *ctx)
+static inline size_t *meta_head(const ag_memblock *ctx)
 {
         return &((size_t *)ctx)[-2];
 }
 
 
-static inline size_t meta_sz(const ag_mblock *ctx)
+static inline size_t meta_sz(const ag_memblock *ctx)
 {
         return ((size_t *)ctx)[-1];
 }
 
 
-static inline size_t meta_refc(const ag_mblock *ctx)
+static inline size_t meta_refc(const ag_memblock *ctx)
 {
         return ((size_t *)ctx)[-2];
 }
 
 
-extern void ag_mblock_exception_handler(const struct ag_exception *ex,
+extern void ag_memblock_exception_handler(const struct ag_exception *ex,
                 void *opt)
 {
-        struct ag_mblock_exception *x = (struct ag_mblock_exception *) opt;
+        struct ag_memblock_exception *x = (struct ag_memblock_exception *) opt;
 
         printf("[!] %d [%s(), %s:%lu]: %s\n", ex->erno, ex->func, ex->file,
                         ex->line, ag_exception_registry_msg(ex->erno));
@@ -56,11 +56,11 @@ extern void ag_mblock_exception_handler(const struct ag_exception *ex,
 }
 
 
-extern ag_mblock *ag_mblock_new(size_t sz)
+extern ag_memblock *ag_memblock_new(size_t sz)
 {
         AG_ASSERT (is_size_valid(sz));
 
-        struct ag_mblock_exception x = {
+        struct ag_memblock_exception x = {
                 .sz = sz,
                 .align = 0,
         };
@@ -72,16 +72,16 @@ extern ag_mblock *ag_mblock_new(size_t sz)
         ctx[0] = 1;
         ctx[1] = sz;
 
-        return (ag_mblock *) &(ctx[2]);
+        return (ag_memblock *) &(ctx[2]);
 }
 
 
-extern ag_mblock *ag_mblock_new_align(size_t sz, size_t align)
+extern ag_memblock *ag_memblock_new_align(size_t sz, size_t align)
 {
         AG_ASSERT (is_size_valid(sz));
         AG_ASSERT (is_alignment_valid(align));
         
-        struct ag_mblock_exception x = {
+        struct ag_memblock_exception x = {
                 .sz = sz,
                 .align = align,
         };
@@ -94,45 +94,45 @@ extern ag_mblock *ag_mblock_new_align(size_t sz, size_t align)
         ctx[0] = 1;
         ctx[1] = sz;
 
-        return (ag_mblock *)&(ctx[2]);
+        return (ag_memblock *)&(ctx[2]);
 }
 
 
-extern ag_mblock *ag_mblock_copy(const ag_mblock *ctx)
+extern ag_memblock *ag_memblock_copy(const ag_memblock *ctx)
 {
-        ag_mblock *cp = (ag_mblock *)ctx;
+        ag_memblock *cp = (ag_memblock *)ctx;
         ((size_t *)cp)[-2]++;
 
         return cp;
 }
 
 
-extern ag_mblock *ag_mblock_clone(const ag_mblock *ctx)
+extern ag_memblock *ag_memblock_clone(const ag_memblock *ctx)
 {
         AG_ASSERT_PTR (ctx);
 
         size_t sz = meta_sz(ctx);
-        ag_mblock *cp = ag_mblock_new(sz);
+        ag_memblock *cp = ag_memblock_new(sz);
         memcpy(cp, ctx, sz);
 
         return cp;
 }
 
 
-extern ag_mblock *ag_mblock_clone_align(const ag_mblock *ctx, size_t align)
+extern ag_memblock *ag_memblock_clone_align(const ag_memblock *ctx, size_t align)
 {
         AG_ASSERT_PTR (ctx);
         AG_ASSERT (is_alignment_valid(align));
 
         size_t sz = meta_sz(ctx);
-        ag_mblock *cp = ag_mblock_new_align(sz, align);
+        ag_memblock *cp = ag_memblock_new_align(sz, align);
         memcpy(cp, ctx, sz);
 
         return cp;
 }
 
 
-extern void ag_mblock_release(ag_mblock **ctx)
+extern void ag_memblock_release(ag_memblock **ctx)
 {
         size_t *hnd;
 
@@ -144,26 +144,26 @@ extern void ag_mblock_release(ag_mblock **ctx)
         }
 }
 
-extern enum ag_cmp ag_mblock_cmp(const ag_mblock *ctx, const ag_mblock *cmp)
+extern enum ag_cmp ag_memblock_cmp(const ag_memblock *ctx, const ag_memblock *cmp)
 {
         return ctx == cmp ? AG_CMP_EQ : memcmp(ctx, cmp, meta_sz(ctx));
 }
 
-extern inline bool ag_mblock_lt(const ag_mblock *, const ag_mblock *);
+extern inline bool ag_memblock_lt(const ag_memblock *, const ag_memblock *);
 
-extern inline bool ag_mblock_eq(const ag_mblock *, const ag_mblock *);
+extern inline bool ag_memblock_eq(const ag_memblock *, const ag_memblock *);
 
-extern inline bool ag_mblock_gt(const ag_mblock *, const ag_mblock *);
+extern inline bool ag_memblock_gt(const ag_memblock *, const ag_memblock *);
 
 
-extern size_t ag_mblock_sz(const ag_mblock *ctx)
+extern size_t ag_memblock_sz(const ag_memblock *ctx)
 {
         AG_ASSERT_PTR (ctx);
 
         return meta_sz(ctx);
 }
 
-extern size_t ag_mblock_sz_total(const ag_mblock *ctx)
+extern size_t ag_memblock_sz_total(const ag_memblock *ctx)
 {
         AG_ASSERT_PTR (ctx);
 
@@ -173,7 +173,7 @@ extern size_t ag_mblock_sz_total(const ag_mblock *ctx)
 
 
 
-extern size_t ag_mblock_refc(const ag_mblock *ctx)
+extern size_t ag_memblock_refc(const ag_memblock *ctx)
 {
         AG_ASSERT_PTR (ctx);
 
@@ -181,7 +181,7 @@ extern size_t ag_mblock_refc(const ag_mblock *ctx)
 }
 
 
-extern bool ag_mblock_aligned(const ag_mblock *ctx, size_t align)
+extern bool ag_memblock_aligned(const ag_memblock *ctx, size_t align)
 {
         AG_ASSERT_PTR (ctx);
         AG_ASSERT (is_alignment_valid(align));
@@ -190,47 +190,47 @@ extern bool ag_mblock_aligned(const ag_mblock *ctx, size_t align)
 }
 
 
-extern void ag_mblock_resize(ag_mblock **ctx, size_t sz)
+extern void ag_memblock_resize(ag_memblock **ctx, size_t sz)
 {
         AG_ASSERT_PTR (ctx && *ctx);
         AG_ASSERT (is_size_valid(sz));
 
-        ag_mblock *hnd = *ctx;
+        ag_memblock *hnd = *ctx;
         size_t oldsz = meta_sz(hnd);
 
-        ag_mblock *cp = ag_mblock_new(sz);
+        ag_memblock *cp = ag_memblock_new(sz);
         memcpy(cp, hnd, sz < oldsz ? sz : oldsz);
         
-        ag_mblock_release(ctx);
+        ag_memblock_release(ctx);
         *ctx = cp;
 }
 
 
-extern void ag_mblock_resize_align(ag_mblock **ctx, size_t sz, size_t align)
+extern void ag_memblock_resize_align(ag_memblock **ctx, size_t sz, size_t align)
 {
         AG_ASSERT_PTR (ctx && *ctx);
         AG_ASSERT (is_size_valid(sz));
         AG_ASSERT (is_alignment_valid(align));
 
-        ag_mblock *hnd = *ctx;
+        ag_memblock *hnd = *ctx;
         size_t oldsz = meta_sz(hnd);
 
-        ag_mblock *cp = ag_mblock_new_align(sz, align);
+        ag_memblock *cp = ag_memblock_new_align(sz, align);
         memcpy(cp, hnd, sz < oldsz ? sz : oldsz);
         
-        ag_mblock_release(ctx);
+        ag_memblock_release(ctx);
         *ctx = cp;
 }
 
 
-extern ag_string *ag_mblock_str(const ag_mblock *ctx)
+extern ag_string *ag_memblock_str(const ag_memblock *ctx)
 {
         AG_ASSERT_PTR (ctx);
 
         return ag_string_new_fmt("address = %p, data sz = %lu,"
                                  " total data = %lu, refc = %lu",
                                  (void *)meta_head(ctx), meta_sz(ctx),
-                                 ag_mblock_sz_total(ctx), meta_refc(ctx));
+                                 ag_memblock_sz_total(ctx), meta_refc(ctx));
 
 }
